feat: 게시글, 방명록 삭제시 작성유저만 삭제되도록 권한 + 프론트 새로 빌드된파일

+++ src/services/post.js
@@ -30,7 +30,13 @@ module.exports = {
   },
 
   // 게시글 삭제
-  delete: async (id) => {
+  delete: async (id, userId) => {
+    const post = await postRepository.findById(id);
+
+    if (post.userId !== userId) {
+      throw new Error('삭제 권한이 없습니다.');
+    }
+
     const deletedpost = await postRepository.delete(id);
     return deletedpost;
   },
